Issue 51
Added a potential fix to 51 for plot_inducer_characterization. I would like to get some advice on my approach. Basically, I do not call ZeroOnLog unless I obtain a min_value of less than or equal to 0. The tests I have do not seem to have a min_value that small, so I can not test out the results. I found a way to make the ZeroOnLog plots look nicer, but I am still not sure what the zero and start should be. For now, they will start at 0 and go to the smallest non_zero value.